Fix build with freetype>=2.5.4

+++ src/libOpenImageIO/imagebufalgo_draw.cpp
@@ -182,9 +182,9 @@ ImageBufAlgo::render_text (ImageBuf &R, int x, int y, string_view text,
         if (error)
             continue;  // ignore errors
         // now, draw to our target surface
-        for (int j = 0;  j < slot->bitmap.rows; ++j) {
+        for (int j = 0;  j < static_cast<int>(slot->bitmap.rows); ++j) {
             int ry = y + j - slot->bitmap_top;
-            for (int i = 0;  i < slot->bitmap.width; ++i) {
+            for (int i = 0;  i < static_cast<int>(slot->bitmap.width); ++i) {
                 int rx = x + i + slot->bitmap_left;
                 float b = slot->bitmap.buffer[slot->bitmap.pitch*j+i] / 255.0f;
                 R.getpixel (rx, ry, pixelcolor);
